[Quantization] Fix bug

+++ src/operator/quantization/quantized_conv2d-inl.h
@@ -113,7 +113,7 @@ class QuantizedConv2DProp : public OperatorProperty {
     }
 
     out_type->clear();
-    out_type->push_back(mshadow::kInt32);
+    out_type->push_back(mshadow::kFloat32);
     out_type->push_back(mshadow::kFloat32);
     out_type->push_back(mshadow::kFloat32);
     return true;
